move audio module to file
